AI_Project_Old-main/requirements.txt
langchain
langchain-community
langchain-openai
langgraph
openai
faiss-cpu
pydantic
python-dotenv
beautifulsoup4
requests
streamlit
wikipedia
pypdf